requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
tensorflow==2.7.0
Pillow==8.4.0
numpy==1.21.4
pydantic==1.8.2
aiofiles==0.8.0
